Fixed readme and fixed import issues
