+++ ciphey/__init__.py
@@ -4,7 +4,7 @@ from . import iface
 
 from . import basemods
 
-from . import __main__
+from . import ciphey 
 
 if __name__ == '__main__':
-    __main__.main()
+    ciphey.main()
